src/requirements.txt
# Dependencias de verdad
pyTelegramBotAPI==4.16.1
python-dotenv==1.0.0
